Add support for java.nio.file package (#47)
* Wrapper for java.nio.file.

Most of the functionality is wrapped. Some APIs that aren't
friendly to functional effects are not wrapped, such as the file
copy methods using Java input and output streams. Perhaps a reasonable
wrapper API can be found for these in future.

Very little in the way of tests yet.

The silencer plugin is used to silence warnings about JavaConverters
being deprecated, so we can support 2.13 without tedious indirection.

* Fix docs.

* PathSuite -> PathSpec.